% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq_lineartrend.R
\name{chisq_lineartrend}
\alias{chisq_lineartrend}
\title{Chi-Squared Test for No Linear Trend (Dose-Response Analysis)}
\usage{
chisq_lineartrend(tbl)
}
\arguments{
\item{tbl}{Results from a dose-response study.}
}
\description{
Return the test statistic and p-value for a chi-squared test for
no linear trend.
}
\references{
https://online.stat.psu.edu/stat504/book/export/html/710
}
