% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oddsratio_matched.R
\name{oddsratio_matched}
\alias{oddsratio_matched}
\title{Odd Ratio 1-1 (Matched Case-Control Study)}
\usage{
oddsratio_matched(obs, alpha = 0.05)
}
\arguments{
\item{obs}{Results from matched case-control study}

\item{alpha}{Significance level for the confidence interval.}
}
\description{
Return the point estimate, estimated standard error, and
(1-alpha)\% confidence interval for a 1-1 matched case-control study.
}
