% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R249-package.R
\docType{package}
\name{R249-package}
\alias{R249}
\alias{R249-package}
\title{R249: An R Implementation of the Open University Module M249}
\description{
Provides functionality in R to replicate the work in M249. The package is under constant development, so should be considered unstable.
}
\keyword{internal}
