% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskratio.R
\name{riskratio}
\alias{riskratio}
\title{Risk ratio}
\usage{
riskratio(obs, alpha = 0.05)
}
\arguments{
\item{obs}{An r by 2 matrix containing the integer counts of the results.}

\item{alpha}{Significance level for the confidence interval.}
}
\value{
An r by 4 matrix, one row for each exposure level
}
\description{
Return the point estimate, estimated standard error, and
(1-alpha)\% confidence interval for the risk ratio of each exposure level.
}
\details{
The reference exposure is expected in the first row, and either
"no disease" or cases in the first column.
See examples.
}
\examples{
## get results
res <- c(201, 76, 215, 327)
## set names, optional
nam <- c("-", "+")
## declare matrix
obs <- matrix(res, nrow = 2, byrow = TRUE, dimnames = list(nam, nam))
## get risk ratio
riskratio(obs)
## get risk ratio, use 1\% sig level for conf ints
riskratio(obs, alpha = 0.01)
}
